Renaming dir and import statements
