+++ activity/inference/model/metadata.go
